feat: Add image form component to course page
This commit adds a new image form component to the course page. The image form allows instructors to add or edit the course image. It includes functionality for uploading an image and displaying it on the page. When an image is successfully updated, a success toast notification is displayed. If there is an error during the update, an error toast notification is shown.

The commit also imports and includes the newly added image form component in the course page, replacing a commented out version of the component.

Conventional Commits Specification:
feat: A new feature was added